backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
numpy>=1.26,<2.1
pypdf>=5.4,<6.0
faiss-cpu>=1.8,<2.0
scikit-learn>=1.5,<2.0
scipy>=1.13,<2.0
huggingface-hub>=0.36,<1.0
python-dotenv>=1.0,<2.0
